Fun stuff and minor corrections
I was doing some testing with one of the vibe motors and was finding it a bit difficult to hold onto properly as well as push a button in a browser via a track pad… SO! I added something that made life a bit easier.

If you’re running this in Chrome (which you should be!!!) Click on the skeleton’s mouth.

Signed-off-by: Matthew Potter <mdpotter@gmail.com>